+++ CHANGELOG.md
@@ -14,6 +14,7 @@ full changeset diff at the end of each section.
 
 Current Trunk
 -------------
+- Add support for non-nullable locals in wasm GC.
 - Add a new flag to Directize, `--pass-arg=directize-initial-contents-immutable`
   which indicates the initial table contents are immutable. That is the case for
   LLVM, for example, and it allows us to optimize more indirect calls to direct